+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/compliance/directreview/DirectReviewUpdateEmailService.java
@@ -5,22 +5,27 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
+import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.ObjectUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Component;
 
 import gov.healthit.chpl.domain.CertifiedProductSearchDetails;
 import gov.healthit.chpl.domain.Developer;
 import gov.healthit.chpl.domain.compliance.DirectReview;
 import gov.healthit.chpl.email.ChplEmailFactory;
+import gov.healthit.chpl.email.ChplHtmlEmailBuilder;
 import gov.healthit.chpl.exception.EmailNotSentException;
+import gov.healthit.chpl.util.Util;
 
 @Component("directReviewUpdateEmailService")
 public class DirectReviewUpdateEmailService {
+    public static final String CONTEXT_PRODUCT_OWNERSHIP = "PRODUCT_OWNERSHIP";
+    public static final String CONTEXT_DEVELOPER_SPLIT = "DEVELOPER_SPLIT";
+    public static final String CONTEXT_DEVELOPER_JOIN = "DEVELOPER_JOIN";
 
     @Value("${directReview.chplChanges.email}")
     private String chplChangesEmailAddress;
@@ -35,28 +40,24 @@ public class DirectReviewUpdateEmailService {
     private String unknownChangesEmailSubject;
 
     private DirectReviewCachingService directReviewService;
-    private Environment env;
     private ChplEmailFactory chplEmailFactory;
-
+    private ChplHtmlEmailBuilder chplHtmlEmailBuilder;
 
     @Autowired
-    public DirectReviewUpdateEmailService(DirectReviewCachingService directReviewService, Environment env,
-            ChplEmailFactory chplEmailFactory) {
+    public DirectReviewUpdateEmailService(DirectReviewCachingService directReviewService,
+            ChplEmailFactory chplEmailFactory, ChplHtmlEmailBuilder chplHtmlEmailBuilder) {
         this.directReviewService = directReviewService;
-        this.env = env;
         this.chplEmailFactory = chplEmailFactory;
+        this.chplHtmlEmailBuilder = chplHtmlEmailBuilder;
     }
 
     /**
      * A developer and possibly listings under that developer have changed in some way.
-     * @param originalDevelopers
-     * @param changedDevelopers
-     * @param originalListings
-     * @param changedListings
-     */
+    */
     public void sendEmail(List<Developer> originalDevelopers, List<Developer> changedDevelopers,
             Map<Long, CertifiedProductSearchDetails> originalListings,
             Map<Long, CertifiedProductSearchDetails> changedListings,
+            String activityContext,
             Logger logger) {
         List<DirectReview> originalDeveloperDrs = new ArrayList<DirectReview>();
         for (Developer originalDeveloper : originalDevelopers) {
@@ -71,14 +72,14 @@ public class DirectReviewUpdateEmailService {
         if (originalDeveloperDrs == null) {
             try {
                 sendUnknownDirectReviewEmails(originalDevelopers, changedDevelopers, originalListings,
-                        changedListings, logger);
+                        changedListings, activityContext, logger);
             } catch (Exception ex) {
                 logger.error("Could not send email to Jira team: " + ex.getMessage());
             }
         } else if (originalDeveloperDrs != null && originalDeveloperDrs.size() > 0) {
             try {
                 sendDirectReviewEmails(originalDeveloperDrs, originalDevelopers, changedDevelopers, originalListings,
-                        changedListings, logger);
+                        changedListings, activityContext, logger);
             } catch (Exception ex) {
                 logger.error("Could not send email to Jira team: " + ex.getMessage());
             }
@@ -91,6 +92,7 @@ public class DirectReviewUpdateEmailService {
             List<Developer> originalDevelopers, List<Developer> changedDevelopers,
             Map<Long, CertifiedProductSearchDetails> originalListings,
             Map<Long, CertifiedProductSearchDetails> changedListings,
+            String activityContext,
             Logger logger)
         throws EmailNotSentException {
         logger.info("Sending email about direct reviews potentially needing changes.");
@@ -100,25 +102,34 @@ public class DirectReviewUpdateEmailService {
             recipients = chplChangesEmailAddress.split(",");
         }
 
-        String htmlMessage = formatDeveloperActionHtml(originalDevelopers, changedDevelopers);
-        htmlMessage += "<p>The following direct reviews associated with the original developer"
+        String developerList = "The following direct reviews associated with the original developer"
                 + (originalDevelopers.size() == 1 ? " " : "s ")
                 + "may need updates: "
                 + "<ul>";
         for (DirectReview dr : drs) {
-            htmlMessage += String.format("<li>%s</li>", dr.getJiraKey());
+            developerList += String.format("<li>%s</li>", dr.getJiraKey());
         }
-        htmlMessage += "</ul></p>";
+        developerList += "</ul>";
 
-        String chplProductNumberChangedHtml = formatChplProductNumbersChangedHtml(originalListings, changedListings);
-        if (StringUtils.isNotEmpty(chplProductNumberChangedHtml)) {
-            htmlMessage += "Any direct reviews with the following developer-associated listings "
-            + "may require updates due to changes in the CHPL Product Number: "
-            + "<ul>" + chplProductNumberChangedHtml + "</ul>";
+        String chplProductNumbersChangedHtml = "";
+        String chplProductNumbersChangedList = formatChplProductNumbersChangedHtml(originalListings, changedListings);
+        if (StringUtils.isNotEmpty(chplProductNumbersChangedList)) {
+            chplProductNumbersChangedHtml = "Additionally, any direct reviews with the following "
+                    + "developer-associated listings may require updates due to changes "
+                    + "in the CHPL Product Number: "
+            + "<ul>" + chplProductNumbersChangedList + "</ul>";
         }
+
+        String htmlMessage = chplHtmlEmailBuilder.initialize()
+                .heading(chplChangesEmailSubject)
+                .paragraph(null, formatDeveloperActionSummary(originalDevelopers, changedDevelopers, activityContext))
+                .paragraph(null, developerList)
+                .paragraph(null, chplProductNumbersChangedHtml)
+                .footer(false)
+                .build();
 
         chplEmailFactory.emailBuilder().recipients(recipients)
-                .subject(env.getProperty("directReview.chplChanges.emailSubject"))
+                .subject(chplChangesEmailSubject)
                 .htmlMessage(htmlMessage)
                 .sendEmail();
     }
@@ -127,72 +138,82 @@ public class DirectReviewUpdateEmailService {
             List<Developer> changedDevelopers,
             Map<Long, CertifiedProductSearchDetails> originalListings,
             Map<Long, CertifiedProductSearchDetails> changedListings,
+            String activityContext,
             Logger logger) throws EmailNotSentException {
         logger.info("Sending email about unknown direct reviews.");
 
         String[] recipients = new String[] {};
-        String emailAddressProperty = env.getProperty("directReview.unknownChanges.email");
+        String emailAddressProperty = unknownChangesEmailAddress;
         if (!StringUtils.isEmpty(emailAddressProperty)) {
             recipients = emailAddressProperty.split(",");
         }
 
-        String htmlMessage = formatDeveloperActionHtml(originalDevelopers, changedDevelopers);
-        htmlMessage += "<p>Any direct reviews related to the original developers may need updated.</p>";
-
-        String chplProductNumberChangedHtml = formatChplProductNumbersChangedHtml(originalListings, changedListings);
-        if (StringUtils.isNotEmpty(chplProductNumberChangedHtml)) {
-            htmlMessage += "<p>Additionally, any direct reviews with the following "
+        String chplProductNumbersChangedHtml = "";
+        String chplProductNumbersChangedList = formatChplProductNumbersChangedHtml(originalListings, changedListings);
+        if (StringUtils.isNotEmpty(chplProductNumbersChangedList)) {
+            chplProductNumbersChangedHtml = "Additionally, any direct reviews with the following "
                     + "developer-associated listings may require updates due to changes "
                     + "in the CHPL Product Number: "
-            + "<ul>" + chplProductNumberChangedHtml + "</ul>";
+            + "<ul>" + chplProductNumbersChangedList + "</ul>";
         }
 
+        String htmlMessage = chplHtmlEmailBuilder.initialize()
+                .heading(unknownChangesEmailSubject)
+                .paragraph(null, formatDeveloperActionSummary(originalDevelopers, changedDevelopers, activityContext))
+                .paragraph(null, "Any direct reviews related to the original developers may need updated.")
+                .paragraph(null, chplProductNumbersChangedHtml)
+                .footer(false)
+                .build();
+
         chplEmailFactory.emailBuilder().recipients(recipients)
-                .subject(env.getProperty("directReview.unknownChanges.emailSubject"))
+                .subject(unknownChangesEmailSubject)
                 .htmlMessage(htmlMessage)
                 .sendEmail();
     }
 
-    private String formatDeveloperActionHtml(List<Developer> originalDevelopers,
-            List<Developer> changedDevelopers) {
+    private String formatDeveloperActionSummary(List<Developer> originalDevelopers,
+            List<Developer> changedDevelopers, String activityContext) {
         String html = "";
-        if (!ObjectUtils.allNotNull(originalDevelopers, changedDevelopers)) {
+        if (!ObjectUtils.allNotNull(originalDevelopers, changedDevelopers)
+                || CollectionUtils.isEmpty(originalDevelopers)
+                || CollectionUtils.isEmpty(changedDevelopers)
+                || StringUtils.isEmpty(activityContext)) {
             return html;
-        } else if (originalDevelopers.size() == 1 && changedDevelopers.size() > 1) {
+        } else if (CONTEXT_DEVELOPER_SPLIT.equals(activityContext)) {
             Developer newDeveloper = null;
             for (Developer changedDeveloper : changedDevelopers) {
                 if (!(changedDeveloper.getId().equals(originalDevelopers.get(0).getId()))) {
                     newDeveloper = changedDeveloper;
                 }
             }
-            html = String.format("<p>The developer %s (ID: %s) was split. "
-                    + "The new developer is %s (ID: %s).</p>",
+            html = String.format("The developer %s (ID: %s) was split. "
+                    + "The new developer is %s (ID: %s).",
                     originalDevelopers.get(0).getName(),
                     originalDevelopers.get(0).getId(),
                     newDeveloper.getName(),
                     newDeveloper.getId());
-        }  else if (originalDevelopers.size() > 1 && changedDevelopers.size() == 1) {
+        }  else if (CONTEXT_DEVELOPER_JOIN.equals(activityContext)) {
             List<String> originalDeveloperNames = originalDevelopers.stream()
                     .map(dev -> dev.getName())
                     .collect(Collectors.toList());
             List<String> originalDeveloperIds = originalDevelopers.stream()
                     .map(dev -> dev.getId().toString())
                     .collect(Collectors.toList());
-            html = String.format("<p>The developers %s (IDs: %s) were merged into a single new developer. "
-                    + "The newly created developer is %s (ID: %s).</p>",
-                    String.join(",", originalDeveloperNames),
-                    String.join(",", originalDeveloperIds),
+            html = String.format("The developer%s %s (IDs: %s) joined the developer %s (ID: %s).",
+                    (originalDeveloperNames.size() > 1 ? "s" : ""),
+                    Util.joinListGrammatically(originalDeveloperNames),
+                    Util.joinListGrammatically(originalDeveloperIds),
                     changedDevelopers.get(0).getName(),
                     changedDevelopers.get(0).getId());
-        } else if (originalDevelopers.size() == 1 && changedDevelopers.size() == 1) {
-            html = String.format("<p>A product changed ownership from developer %s (IDs: %s) "
-                    + "to developer %s (ID: %s).</p>",
+        } else if (CONTEXT_PRODUCT_OWNERSHIP.equals(activityContext)) {
+            html = String.format("A product changed ownership from developer %s (IDs: %s) "
+                    + "to developer %s (ID: %s).",
                     originalDevelopers.get(0).getName(),
                     originalDevelopers.get(0).getId(),
                     changedDevelopers.get(0).getName(),
                     changedDevelopers.get(0).getId());
         } else {
-            html += "<p>A change to developers occurred.</p>";
+            html += "A change to developers occurred.";
         }
         return html;
     }