+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/permissions/domains/DeveloperDomainPermissions.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Component;
 import gov.healthit.chpl.permissions.domains.developer.CreateActionPermissions;
 import gov.healthit.chpl.permissions.domains.developer.GetAllUsersActionPermissions;
 import gov.healthit.chpl.permissions.domains.developer.GetAllWithDeletedActionPermissions;
-import gov.healthit.chpl.permissions.domains.developer.MergeActionPermissions;
+import gov.healthit.chpl.permissions.domains.developer.JoinActionPermissions;
 import gov.healthit.chpl.permissions.domains.developer.SplitActionPermissions;
 import gov.healthit.chpl.permissions.domains.developer.UpdateActionPermissions;
 
@@ -16,7 +16,7 @@ public class DeveloperDomainPermissions extends DomainPermissions {
     public static final String GET_ALL_WITH_DELETED = "GET_ALL_WITH_DELETED";
     public static final String UPDATE = "UPDATE";
     public static final String CREATE = "CREATE";
-    public static final String MERGE = "MERGE";
+    public static final String JOIN = "JOIN";
     public static final String SPLIT = "SPLIT";
     public static final String GET_ALL_USERS = "GET_ALL_USERS";
 
@@ -25,14 +25,14 @@ public class DeveloperDomainPermissions extends DomainPermissions {
             @Qualifier("developerGetAllWithDeletedActionPermissions") GetAllWithDeletedActionPermissions getAllWithDeletedActionPermissions,
             @Qualifier("developerUpdateActionPermissions") UpdateActionPermissions updateActionPermissions,
             @Qualifier("developerCreateActionPermissions") CreateActionPermissions createActionPermissions,
-            @Qualifier("developerMergeActionPermissions") MergeActionPermissions mergeActionPermissions,
+            @Qualifier("developerJoinActionPermissions") JoinActionPermissions joinActionPermissions,
             @Qualifier("developerSplitActionPermissions") SplitActionPermissions splitActionPermissions,
             @Qualifier("developerGetAllUsersActionPermissions") GetAllUsersActionPermissions getUsersActionPermissions) {
 
         getActionPermissions().put(GET_ALL_WITH_DELETED, getAllWithDeletedActionPermissions);
         getActionPermissions().put(UPDATE, updateActionPermissions);
         getActionPermissions().put(CREATE, createActionPermissions);
-        getActionPermissions().put(MERGE, mergeActionPermissions);
+        getActionPermissions().put(JOIN, joinActionPermissions);
         getActionPermissions().put(SPLIT, splitActionPermissions);
         getActionPermissions().put(GET_ALL_USERS, getUsersActionPermissions);
     }