+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/scheduler/job/onetime/participants/ParticipantUpdateStrategy.java
@@ -1,8 +1,0 @@
-package gov.healthit.chpl.scheduler.job.onetime.participants;
-
-import gov.healthit.chpl.domain.CertifiedProductSearchDetails;
-
-public interface ParticipantUpdateStrategy {
-
-    boolean updateParticipants(CertifiedProductSearchDetails listing);
-}